+++ docs/CHANGELOG.md
@@ -12,6 +12,17 @@ Action Keywords:
 - `FIX`   : Functionality existed but did not behave as expected.
 - `REMOVE` : Functionality is no longer available.
 
+## [0.63.16] - 2025-02-03
+
+## ADD
+
+- `./libs/bash/aws-unset_env_var.sh` to remove set AWS_* ENV VARs, useful when switching accounts in using `aws-sso` and the same terminal session
+- optional `./libs/bash/append_libs_to_sys_path.sh` helper that will append lib paths to system,s ENV VAR PATH
+
+## FIX
+
+- `./libs/jenkins/pipelines/vars/ContainerImage.groovy` now uses `podman ...--squash` argument when building container images to reduce image sizes
+
 ## [0.63.15] - 2025-01-28
 
 ## REMOVE